Backend/requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
flask_session==0.8.0
flask_sqlalchemy==3.1.1
mediapipe==0.10.8
numpy==1.26.4
opencv_contrib_python
Pillow==10.3.0
psutil==5.9.8
redis==5.0.3
tqdm==4.66.2
Werkzeug==3.0.2
